+++ jhdf/src/main/java/io/jhdf/api/Dataset.java
@@ -50,8 +50,17 @@ public interface Dataset extends Node {
 	 * Gets the dimensions of this dataset
 	 *
 	 * @return the dimensions of this dataset
+	 * @throws ArithmeticException if any dimension exceeds {@link Integer#MAX_VALUE}
 	 */
 	int[] getDimensions();
+
+	/**
+	 * Gets the dimensions of this dataset as long values. This supports datasets with dimensions
+	 * exceeding {@link Integer#MAX_VALUE}.
+	 *
+	 * @return the dimensions of this dataset
+	 */
+	long[] getDimensionsAsLong();
 
 	/**
 	 * Checks if this dataset is scalar i.e is a single element with no dimensions.