+++ jhdf/src/main/java/io/jhdf/dataset/ContiguousDatasetImpl.java
@@ -69,7 +69,7 @@ public class ContiguousDatasetImpl extends DatasetBase implements ContiguousData
 			}
 		} else {
 			// fastest dim so read some data
-			long offsetBytes = Utils.dimensionIndexToLinearIndex(currentOffset, getDimensions()) * getDataType().getSize();
+			long offsetBytes = Utils.dimensionIndexToLinearIndex(currentOffset, getDimensionsAsLong()) * getDataType().getSize();
 			long fileOffset = contiguousDataLayoutMessage.getAddress() + offsetBytes;
 
 			final int fastestDimLengthBytes = sliceDimensions[sliceDimensions.length - 1] * getDataType().getSize();